+++ COT/commands/tests/test_add_file.py
@@ -19,7 +19,7 @@
 import os.path
 
 from COT.tests.ut import COT_UT
-from COT.ui_shared import UI
+from COT.ui import UI
 from COT.commands.add_file import COTAddFile
 from COT.data_validation import InvalidInputError