+++ COT/commands/tests/test_edit_hardware.py
@@ -19,7 +19,7 @@
 import re
 
 from COT.tests.ut import COT_UT
-from COT.ui_shared import UI
+from COT.ui import UI
 from COT.commands.edit_hardware import COTEditHardware
 from COT.data_validation import InvalidInputError
 from COT.platforms import IOSv, IOSXRv