+++ COT/commands/tests/test_info.py
@@ -17,7 +17,7 @@
 """Unit test cases for the COT.info.COTInfo class."""
 
 from COT.tests.ut import COT_UT
-from COT.ui_shared import UI
+from COT.ui import UI
 from COT.commands.info import COTInfo
 from COT.data_validation import InvalidInputError