+++ COT/ui/tests/test_cli.py
@@ -3,7 +3,7 @@
 # test_cli.py - Unit test cases for generic COT CLI.
 #
 # September 2013, Glenn F. Matthews
-# Copyright (c) 2013-2016 the COT project developers.
+# Copyright (c) 2013-2017 the COT project developers.
 # See the COPYRIGHT.txt file at the top-level directory of this distribution
 # and at https://github.com/glennmatthews/cot/blob/master/COPYRIGHT.txt.
 #
@@ -14,7 +14,7 @@
 # of COT, including this file, may be copied, modified, propagated, or
 # distributed except according to the terms contained in the LICENSE.txt file.
 
-"""Unit test cases for the COT.cli.CLI class and related functionality."""
+"""Unit test cases for the COT.ui.cli.CLI class and related functionality."""
 
 from __future__ import print_function
 
@@ -32,7 +32,7 @@ import mock
 
 from COT import __version_long__
 from COT.tests.ut import COT_UT
-from COT.cli import CLI
+from COT.ui.cli import CLI
 from COT.data_validation import InvalidInputError
 
 # pylint: disable=missing-param-doc,missing-type-doc